wait for identity permission to be granted before adding identity listeners

+++ src/background/state-managers/identity-state-manager.ts
@@ -1,12 +1,36 @@
 import {OAuth2Service} from '../../app/services/drive-api/o-auth-2.service';
 import {getCurrentTimeStringWithMillis} from '../../app/utils/date-utils';
+import {ChromePermissionsService} from '../../app/services/chrome-permissions.service';
 
 export class IdentityStateManager {
 
-  constructor(private oAuth2Service: OAuth2Service) {
-    chrome.identity.onSignInChanged.addListener(() => {
-      console.log(getCurrentTimeStringWithMillis(), '- chrome sign in status changed');
-      this.oAuth2Service.removeCachedAuthToken();
+  private signInListenerInitialized = false;
+
+  constructor(private oAuth2Service: OAuth2Service,
+              private chromePermissionsService: ChromePermissionsService) {
+    this.chromePermissionsService.permissionsUpdated$.subscribe(() => {
+      this.setupListeners();
+    });
+    this.setupListeners();
+  }
+
+  private setupListeners() {
+    this.chromePermissionsService.hasIdentityPermissions().then(hasIdentityPermissions => {
+      if (hasIdentityPermissions && !this.signInListenerInitialized) {
+        this.addOnSignInChangedListener();
+        this.signInListenerInitialized = true;
+      }
+    });
+  }
+
+  private addOnSignInChangedListener() {
+    chrome.identity.onSignInChanged.addListener((account, signedIn) => {
+      console.log(`${getCurrentTimeStringWithMillis()} - chrome sign in status changed, signedIn=${signedIn}`);
+      if (signedIn) {
+        this.oAuth2Service.removeCachedAuthToken();
+      } else {
+        this.oAuth2Service.refreshAuthStatus();
+      }
     });
   }
 }
